Create `DashboardViewModel` to sync dashboard data and determine the stats version.
