+++ WooCommerce/Classes/ViewRelated/Dashboard/DashboardViewModel.swift
@@ -1,0 +1,89 @@
+import Yosemite
+import enum Networking.DotcomError
+import enum Storage.StatsVersion
+
+/// Syncs data for dashboard stats UI and determines the state of the dashboard UI based on stats version.
+final class DashboardViewModel {
+    /// Stats v4 is shown by default, then falls back to v3 if store stats are unavailable.
+    @Published private(set) var statsVersion: StatsVersion = .v4
+
+    /// Syncs store stats for dashboard UI.
+    func syncStats(for siteID: Int64,
+                   siteTimezone: TimeZone,
+                   timeRange: StatsTimeRangeV4,
+                   latestDateToInclude: Date,
+                   onCompletion: ((Result<Void, Error>) -> Void)? = nil) {
+        let earliestDateToInclude = timeRange.earliestDate(latestDate: latestDateToInclude, siteTimezone: siteTimezone)
+        let action = StatsActionV4.retrieveStats(siteID: siteID,
+                                                 timeRange: timeRange,
+                                                 earliestDateToInclude: earliestDateToInclude,
+                                                 latestDateToInclude: latestDateToInclude,
+                                                 quantity: timeRange.maxNumberOfIntervals,
+                                                 onCompletion: { [weak self] result in
+            guard let self = self else { return }
+            switch result {
+            case .success:
+                break
+            case .failure(let error):
+                DDLogError("⛔️ Dashboard (Order Stats) — Error synchronizing order stats v4: \(error)")
+                if error as? DotcomError == .noRestRoute {
+                    self.statsVersion = .v3
+                }
+            }
+            onCompletion?(result)
+        })
+        ServiceLocator.stores.dispatch(action)
+    }
+
+    /// Syncs visitor stats for dashboard UI.
+    func syncSiteVisitStats(for siteID: Int64,
+                            siteTimezone: TimeZone,
+                            timeRange: StatsTimeRangeV4,
+                            latestDateToInclude: Date,
+                            onCompletion: ((Result<Void, Error>) -> Void)? = nil) {
+        let action = StatsActionV4.retrieveSiteVisitStats(siteID: siteID,
+                                                          siteTimezone: siteTimezone,
+                                                          timeRange: timeRange,
+                                                          latestDateToInclude: latestDateToInclude,
+                                                          onCompletion: { result in
+            if case let .failure(error) = result {
+                DDLogError("⛔️ Error synchronizing visitor stats: \(error)")
+            }
+            onCompletion?(result)
+        })
+        ServiceLocator.stores.dispatch(action)
+    }
+
+    /// Syncs top performers data for dashboard UI.
+    func syncTopEarnersStats(for siteID: Int64,
+                             siteTimezone: TimeZone,
+                             timeRange: StatsTimeRangeV4,
+                             latestDateToInclude: Date,
+                             onCompletion: ((Result<Void, Error>) -> Void)? = nil) {
+        let earliestDateToInclude = timeRange.earliestDate(latestDate: latestDateToInclude, siteTimezone: siteTimezone)
+        let action = StatsActionV4.retrieveTopEarnerStats(siteID: siteID,
+                                                          timeRange: timeRange,
+                                                          earliestDateToInclude: earliestDateToInclude,
+                                                          latestDateToInclude: latestDateToInclude,
+                                                          quantity: Constants.topEarnerStatsLimit,
+                                                          onCompletion: { result in
+            switch result {
+            case .success:
+                ServiceLocator.analytics.track(event:
+                        .Dashboard.dashboardTopPerformersLoaded(timeRange: timeRange))
+            case .failure(let error):
+                DDLogError("⛔️ Dashboard (Top Performers) — Error synchronizing top earner stats: \(error)")
+            }
+            onCompletion?(result)
+        })
+        ServiceLocator.stores.dispatch(action)
+    }
+}
+
+// MARK: - Constants
+//
+private extension DashboardViewModel {
+    enum Constants {
+        static let topEarnerStatsLimit: Int = 5
+    }
+}
